tidy up, refactor some components, relative imports
