+++ src/components/Tutorial/Assessment.jsx
@@ -12,8 +12,8 @@ import Typography from "@mui/material/Typography";
 import * as Blockly from "blockly";
 import { initialXml } from "../Blockly/initialXml";
 import IconButton from "@mui/material/IconButton";
-import CodeViewer from "../CodeViewer";
-import TooltipViewer from "../TooltipViewer";
+import CodeViewer from "../Workspace/CodeViewer";
+import TooltipViewer from "../Workspace/TooltipViewer";
 import Tooltip from "@mui/material/Tooltip";
 import ReactMarkdown from "react-markdown";
 import { faCode } from "@fortawesome/free-solid-svg-icons";
