+++ src/components/Workspace/ToolbarItems/AutoSave.jsx
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import { workspaceName } from "../../../actions/workspaceActions";
-import SaveIcon from "../../CodeEditor/SaveIcon";
+import SaveIcon from "@/components/Pages/CodeEditor/SaveIcon";
 
 const resetTimeout = (id, newID) => {
   clearTimeout(id);
