+++ vite.config.js
@@ -6,6 +6,11 @@ export default defineConfig(() => {
     build: {
       outDir: "build",
     },
+    resolve: {
+      alias: {
+        "@": "/src",
+      },
+    },
     plugins: [react()],
     server: {
       port: 3000,
